// src/lib/spacetime/dispatch/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const DispatchEvent = __t.object("DispatchEvent", {
  eventId: __t.string(),
  vendorId: __t.i32(),
  routeUuid: __t.string(),
  stopUuid: __t.string(),
  driverUuid: __t.string(),
  eventType: __t.string(),
  status: __t.string(),
  message: __t.string(),
  recordedAt: __t.string(),
});
export type DispatchEvent = __Infer<typeof DispatchEvent>;

export const DriverLocation = __t.object("DriverLocation", {
  driverUuid: __t.string(),
  userUuid: __t.string(),
  vendorId: __t.i32(),
  routeUuid: __t.string(),
  latitude: __t.f64(),
  longitude: __t.f64(),
  speedMps: __t.f64(),
  heading: __t.f64(),
  accuracyMeters: __t.f64(),
  hardBrake: __t.bool(),
  hardBrakeDeltaMps: __t.f64(),
  crashDetected: __t.bool(),
  fallDetected: __t.bool(),
  source: __t.string(),
  recordedAt: __t.string(),
  updatedAt: __t.string(),
});
export type DriverLocation = __Infer<typeof DriverLocation>;

export const DriverLocationSample = __t.object("DriverLocationSample", {
  sampleId: __t.string(),
  driverUuid: __t.string(),
  userUuid: __t.string(),
  vendorId: __t.i32(),
  routeUuid: __t.string(),
  latitude: __t.f64(),
  longitude: __t.f64(),
  speedMps: __t.f64(),
  previousSpeedMps: __t.f64(),
  hardBrake: __t.bool(),
  hardBrakeDeltaMps: __t.f64(),
  crashDetected: __t.bool(),
  fallDetected: __t.bool(),
  source: __t.string(),
  recordedAt: __t.string(),
});
export type DriverLocationSample = __Infer<typeof DriverLocationSample>;